requirements.txt
Flask
gTTS
python-docx
PyPDF2
werkzeug
python-magic-bin ; si estás en Windows.
python-magic ; si no estás en Windows.
gunicorn